fix : 피드백 반영(불필요한 파일 삭제 및 구조 변경)
