+++ src/main/kotlin/lotto/LottoApplication.kt
@@ -1,22 +1,24 @@
 package lotto
 
-import lotto.configuration.LottoConfiguration
-import lotto.controller.LottoController
+import lotto.domain.Amount
+import lotto.domain.LottoGame
+import lotto.domain.LottoGameResult
+import lotto.repository.LottoRepository
 import lotto.view.InputView
 import lotto.view.OutPutView
 
-private val lottoController: LottoController =
-    LottoConfiguration().LottoController()
 
 fun main() {
-    val amount = InputView.getLottoPurchaseAmount()
+    val lottoGame = LottoGame(LottoRepository())
+    val purchaseAmount = InputView.getLottoPurchaseAmount()
     val manualCount = InputView.getLottoManualCount()
-    amount.setManualCount(manualCount)
-    val lottoManualNumbers = InputView.getLottoManualNumbers(amount.getManualCount())
-    val lottoInfo = lottoController.start(amount, lottoManualNumbers)
+    val amount = Amount(purchaseAmount, manualCount)
+
+    val lottoManualNumbers = InputView.getLottoManualNumbers(amount.manualCount)
+    val lottoInfo = lottoGame.start(amount, lottoManualNumbers)
     OutPutView.printLottoInfo(lottoInfo)
 
     val winnerNumbers = InputView.getWinningNumber()
-    val resultResponse = lottoController.getResult(winnerNumbers)
-    OutPutView.printLottoResults(resultResponse, amount.amount)
+    val lottoGameResult = LottoGameResult(lottoInfo, winnerNumbers)
+    OutPutView.printLottoResults(lottoGameResult.getResult(), amount.amount)
 }
